refactor: agno.tools.file and agno.tools.knowledge as packages
Both toolkits that belong together now live under one import path:

    from agno.tools.file import FileGenerationTools
    from agno.tools.knowledge import KnowledgeManagementTools

file.py and knowledge.py become packages holding their sibling toolkit,
following agno.tools.mcp and agno.tools.finance. Names resolve on first
access (PEP 562), so importing FileTools no longer pulls reportlab and
python-docx in behind it (45 ms when both are installed).

agno.tools.file_generation keeps working as a re-export; it has been
public since 2.x. agno.tools.knowledge_management is removed — it
shipped one day ago in v3.0.3.

KnowledgeManagementTools drops the 2.x-style enable_* flags for flags
named after the tools they register:

    ingest_url=True, ingest_path=False, ingest_text=True, remove_content=True

ingest_path is now opt-in: it reads any path the server process can read,
and under scope="shared" what it loads becomes readable by every agent on
that knowledge base. So the common case needs no exclude_tools:

    KnowledgeManagementTools(knowledge=product_knowledge)

Also fixes a fail-open in the same constructor: requires_confirmation_tools
was set with setdefault, so a caller naming any other tool silently dropped
the confirmation gate on remove_content. It now unions.
