+++ cookbook/05_agent_os/02_databases/gcs_media_storage.py
@@ -21,7 +21,7 @@ from agno.db.sqlite import SqliteDb
 from agno.media.storage.gcs import AsyncGCSMediaStorage
 from agno.models.openai import OpenAIResponses
 from agno.os import AgentOS
-from agno.tools.file_generation import FileGenerationTools
+from agno.tools.file import FileGenerationTools
 from dotenv import load_dotenv
 
 load_dotenv()
